completed restructuring code for organization, fixed some warnings
